functions/requirements.txt
firebase_functions
firebase_admin
python-dotenv
openai==1.93.0
# LangChain minimal set used:
langchain-core
langchain-text-splitters
langchain-openai
langchain-pinecone
pinecone
